reformat test scripts, rename src files
